requirements.txt
Flask==3.1.0
gunicorn==21.2.0
Werkzeug==3.1.5
Jinja2==3.1.6
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
cloudinary==1.43.0
python-dotenv==1.0.1